+++ Infrastructure/Service/HomeBuildService.cs
@@ -1,6 +1,7 @@
 ﻿using Aplication.Service;
 using Domain.Entitys.Home.Services;
 using Domain.Model;
+using Infrastructure.DataAccess;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -11,9 +12,22 @@ namespace Infrastructure.Service
 {
     public class HomeBuildService : IHomeBuild
     {
-        public Task<Response<HomeBuild>> CreateHomeBuildAynce(HomeBuild homeBuild)
+        private readonly HomeDbContext _homeDbContext;
+
+       
+
+        public async Task<Response<HomeBuild>> CreateHomeBuildAynce(HomeBuild homeBuild)
         {
-            throw new NotImplementedException();
+            try
+            {
+                await _homeDbContext.HomeBuilds.AddAsync(homeBuild);
+                await _homeDbContext.SaveChangesAsync();
+                return new Response<HomeBuild>(homeBuild);
+            }
+            catch (Exception ex)
+            {
+                return new Response<HomeBuild>("Failed to create homeBuild: " + ex.Message);
+            }
         }
 
         public Task<Response<bool>> DeleteHomeBuildAynce(int Id)
